Split GskModule into GskModule & GskPublishedModule in buildSrc
